+++ spec/flipper/railtie_spec.rb
@@ -3,11 +3,8 @@ require 'flipper/railtie'
 
 RSpec.describe Flipper::Railtie do
   let(:application) do
-    Class.new(Rails::Application).new(
-      railties: [Flipper::Railtie],
-    ).tap do |app|
-      app.config.eager_load = false
-      app.run_load_hooks!
+    Class.new(Rails::Application).create(railties: [Flipper::Railtie]) do
+      config.eager_load = false
     end
   end
 
@@ -22,46 +19,55 @@ RSpec.describe Flipper::Railtie do
 
   describe 'initializers' do
     it 'sets defaults' do
+      subject # initialize
       expect(config.env_key).to eq("flipper")
       expect(config.memoize).to be(true)
       expect(config.preload).to be(true)
     end
 
     it "configures instrumentor on default instance" do
-      subject
-
+      subject # initialize
       expect(Flipper.instance.instrumenter).to eq(ActiveSupport::Notifications)
     end
 
     it 'uses Memoizer middleware if config.memoize = true' do
+      initializer { config.memoize = true }
       expect(subject.middleware).to include(Flipper::Middleware::Memoizer)
     end
 
     it 'does not use Memoizer middleware if config.memoize = false' do
-      # load but don't initialize
-      config.memoize = false
-
+      initializer { config.memoize = false }
       expect(subject.middleware).not_to include(Flipper::Middleware::Memoizer)
     end
 
     it 'passes config to memoizer' do
-      # load but don't initialize
-      config.update(
-        env_key: 'my_flipper',
-        preload: [:stats, :search]
-      )
+      initializer do
+        config.update(
+          env_key: 'my_flipper',
+          preload: [:stats, :search]
+        )
+      end
 
-      expect(Flipper::Middleware::Memoizer).to receive(:new).with(application.routes,
-        env_key: 'my_flipper', preload: [:stats, :search], if: nil
-      )
-
-      subject # initialize
+      expect(subject.middleware).to include(Flipper::Middleware::Memoizer)
+      middleware = subject.middleware.detect { |m| m.klass == Flipper::Middleware::Memoizer }
+      expect(middleware.args[0]).to eq({
+        env_key: config.env_key,
+        preload: config.preload,
+        if: nil
+      })
     end
 
     it "defines #flipper_id on AR::Base" do
       subject
       require 'active_record'
       expect(ActiveRecord::Base.ancestors).to include(Flipper::Identifier)
+    end
+  end
+
+  # Add app initializer in the same order as config/initializers/*
+  def initializer(&block)
+    application.initializer 'spec', before: :load_config_initializers do
+      block.call
     end
   end
 end